refactor: 기존 디바이스의 Orientation을 설정을 LockOperation 프로토콜의 기본 구현으로 분리
